api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
bcrypt==4.0.1
passlib==1.7.4
